UseCases/Enterprise_Vector_Store/chat_app/requirements.txt
teradataml>=20.0.0.5
teradatagenai>=20.0.0.1
streamlit
python-dotenv
plotly